Update README.md fixing architecture error
* Objective

This commit objective is about fixing and architecture usage in the
README.md examples as it was using the store directly in the job,
instead of using the usecase (or core).

* Why

It was a mistake to initialize one job that used the store layer
directly, it was meant to be using the usercase or core layers, this
commit fixed that.

* How

This commit updates the examples so they now use the use case layer
instead than the store layer.