Update readme including new specs
* Objective

This commit objective is about reflecting on the readme of the new specs
for jobs that we have decided on in a sync meeting.

* Why

Still, this README.md is meant to be a guide while implementing new
jobs, so we need to keep it updated with the latest specs.

* How
This commit is just a README.md updater.